% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd.calculations.R
\name{bd.probability.extant}
\alias{bd.probability.extant}
\title{Birth-death probability (Stadler, 2010)}
\usage{
bd.probability.extant(tree, b, d, rho = 1, mpfr = F)
}
\arguments{
\item{tree}{Phylo object of extant taxa only}

\item{b}{Rate of speciation (branching)}

\item{d}{Rate of extinction (branch termination)}

\item{rho}{Extant species sampling}
}
\value{
Log likelihood
}
\description{
Birth-death probability (Stadler, 2010)
}

