% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd.calculations.R
\name{bd.probability}
\alias{bd.probability}
\title{Birth-death probability (Keiding, 1975)}
\usage{
bd.probability(frs, lambda, mu)
}
\arguments{
\item{frs}{Dataframe of species ranges}

\item{lambda}{Rate of speciation}

\item{mu}{Rate of extinction}
}
\description{
This is the underlying birth-death process used in Silvestro et al. 2014, eq. 9
}

