% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.calculations.pbdb.R
\name{boundary.crosser.rates.pa}
\alias{boundary.crosser.rates.pa}
\title{Calculate speciation and extinction rates using the boundary crosser approach}
\usage{
boundary.crosser.rates.pa(presence.absence, intervals, continuous = TRUE)
}
\arguments{
\item{presence.absence}{Dataframe of presence/absence data. Each row represents a taxon and each column represents a time interval from oldest to youngest.}

\item{intervals}{Dataframe of geological intervals.}

\item{continuous}{If TRUE calculate continuous rates (i.e. account for interval length)}
}
\value{
Dataframe containing the following columns:
\itemize{
\item \code{int} interval number. Intervals are numbered from oldest (=1) to youngest (>1)
\item \code{int.name} Interval name
\item \code{NFl} singletons
\item \code{NFt} top boundary crossers
\item \code{Nbl} bottom boundary crossers
\item \code{Nbt} range through taxa
\item \code{p} speciation rate
\item \code{q} extinction rate
}
Note this approach does not return rates for the first interval.
}
\description{
Calculate speciation and extinction rates using the boundary crosser approach
}
\examples{
set.seed(111)
# simulate tree
birth = 0.02
death = 0.01
tips = 500
t = TreeSim::sim.bd.taxa(tips, 1, birth, death)[[1]]

# simulate fossils
f = FossilSim::sim.fossils.poisson(t, 1)

# add extant occurrences
f = FossilSim::add.extant.occ(t, f)

# asymmetric species assignment (this may take a while)
f = FossilSim::asymmetric.fossil.mapping(t, f)

# generate presence.absence matrix
pa = presence.absence.matrix(f, stages)

# calculate speciation & extinction rates
out = boundary.crosser.rates.pa(pa, stages)

# rates
mean(out$p, na.rm = TRUE)
mean(out$q, na.rm = TRUE)

# plot the output
max = length(stages$name)

plot(1:max, out$p, pch = 16, bty = "n", xlab = "stage #", ylab = "speciation")
lines(1:max, rep(0.02, max), lty = 2, lwd = "2", col = 2)

plot(1:max, out$p, pch = 16, bty = "n", xlab = "stage #", ylab = "extinction")
lines(1:max, rep(0.02, max), lty = 2, lwd = "2", col = "deepskyblue")

}
\references{
Foote, M. 2000. Origination and extinction components of taxonomic diversity: General problems. Paleobiology 26: 74-102.
}
\seealso{
\code{\link{interval.types.bc.pa}}
}
