% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd.likelihood.R
\name{est.bd.extant}
\alias{est.bd.extant}
\title{Maximum likelihood estimation of birth and death rates for a given phylogeny of extant taxa}
\usage{
est.bd.extant(b, d, tree, lower.b = 0.001, upper.b = 10, lower.d = 0.001,
  upper.d = 10, crown = FALSE, rho = 1)
}
\arguments{
\item{b}{Initial value for birth rate}

\item{d}{Initial value for death rate. Must be < b}

\item{tree}{Phylo object of extant taxa (the function will remove any extinct taxa prior to calculating the likelihood)}

\item{lower.b}{Lower birth rate}

\item{upper.b}{Upper birth rate}

\item{lower.d}{Lower death rate}

\item{upper.d}{Upper death rate}

\item{crown}{If TRUE the process is conditioned on the crown (default = F)}

\item{rho}{Extant species sampling probability (default = 1)}
}
\value{
Named list including maximum likelihood estimates of birth and death rate
}
\description{
This function maximises the likelihood of the function bd.probability.extant using optim (Nelder-Mead option)
}
\examples{
# simulate tree
t = TreeSim::sim.bd.taxa(100,1,1,0.1)[[1]]
# estimate birth & death rates
birth = runif(1)
death = runif(1, 0, birth)
out = est.bd.extant(birth, death, t)
# ML birth rate
out$par[1]
# ML death rate
out$par[2]
}

