% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd.likelihood.R
\name{est.bd.ranges}
\alias{est.bd.ranges}
\title{Maximum likelihood estimation of birth and death rates for a given set of stratigraphic ranges}
\usage{
est.bd.ranges(b, d, frs, lower.b = 0.001, upper.b = 10, lower.d = 0.001,
  upper.d = 10, crown = FALSE)
}
\arguments{
\item{b}{Initial value for birth rate}

\item{d}{Initial value for death rate}

\item{frs}{Dataframe of species ranges}

\item{lower.b}{Lower birth rate}

\item{upper.b}{Upper birth rate}

\item{lower.d}{Lower death rate}

\item{upper.d}{Upper death rate}

\item{crown}{If TRUE assume the process begins at the first speciation event and not the origin (default = FALSE)}
}
\value{
Named list including maximum likelihood estimates of birth and death rate
}
\description{
This function maximises the likelihood of the function bd.probability.range using optim (Nelder-Mead option)
}
\examples{
# simulate tree & assume complete sampling
t = TreeSim::sim.bd.taxa(100,1,1,0.1)[[1]]
# add symmetric speciation events & generate fossil range dataframe
f = 0.5
ages <- FossilSim::mixed.ages(t, f, root.edge = TRUE)
# add anagenic speciation events
lambda.a = 0.1
frs <- FossilSim::anagenic.species(ages, lambda.a)
# estimate birth & death rates
birth = runif(1)
death = runif(1)
out = est.bd.ranges(birth, death, frs)
# ML birth rate
out$par[1]
# ML death rate
out$par[2]
}

