% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbd.skyline.calculations.R
\name{fbdskyline.probability}
\alias{fbdskyline.probability}
\alias{reassign.intervals}
\alias{recount.k}
\title{FBD skyline probability}
\usage{
fbdskyline.probability(frs, b, d, s, rho, k, int.min, mpfr = F)

recount.k(fossils, intervals)

reassign.intervals(frs, intervals)
}
\description{
WARNING: FUNCTION RETURNS THE INCORRECT PROBABILITY
}

