% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.calculations.intervals.R
\name{first.last.appearances.pa}
\alias{first.last.appearances.pa}
\title{Work out first and last appearance intervals}
\usage{
first.last.appearances.pa(presence.absence)
}
\arguments{
\item{presence.absence}{Dataframe of presence/absence data. Each row represents a taxon and each column represents a time interval from oldest to youngest.}
}
\value{
Dataframe containing first and last appearance intervals. Intervals are numbered oldest (= 1) to youngest (> 1).
}
\description{
Work out first and last appearance intervals
}
\examples{
set.seed(123)
# simulate tree
t = TreeSim::sim.bd.taxa(100,1,0.01,0.005)[[1]]

# simulate species
sp = FossilSim::sim.taxonomy(t)

# simulate fossils
f = FossilSim::sim.fossils.poisson(1, taxonomy = sp)
f = FossilSim::sim.extant.samples(f, taxonomy = sp)

# generate presence.absence matrix
pa = presence.absence.matrix(f, stages)

# work out first and last appearance intervals
fa = first.last.appearances.pa(pa)

}
