% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.calculations.R
\name{first.last.appearances}
\alias{first.last.appearances}
\title{Work out first and last appearances}
\usage{
first.last.appearances(fossils)
}
\arguments{
\item{fossils}{Dataframe of sampled fossils (sp = edge labels. h = ages.)}
}
\value{
Dataframe containing the first and last appearances for each lineage, represented by the max secure age of the sampling horizon.
}
\description{
Work out first and last appearances
}
\examples{
# simulate tree
t = ape::rtree(10)
# simulate fossils
f = FossilSim::sim.fossils.poisson(1, t)
# work out first & last appearances
first.last.appearances(f)
}
