% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.calculations.R
\name{gap.filler.rates}
\alias{gap.filler.rates}
\title{Calculate speciation and extinction rates using the gap-filler approach}
\usage{
gap.filler.rates(fossils, basin.age, strata)
}
\arguments{
\item{fossils}{Dataframe of sampled fossils (sp = edge labels. h = ages.)}

\item{basin.age}{Maximum age of the oldest stratigraphic interval}

\item{strata}{Number of stratigraphic horizons}

\item{continuous}{If TRUE calculate continuous rates}
}
\description{
Calculate speciation and extinction rates using the gap-filler approach
}

