% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.calculations.tree.wide.R
\name{interval.types.3t}
\alias{interval.types.3t}
\title{Categorise per interval taxon types using the three-timer approach (Alroy, 2008)}
\usage{
interval.types.3t(fossils, max.age, strata, return.useful = FALSE)
}
\arguments{
\item{fossils}{Dataframe of sampled fossils (sp = unique species ID. h = ages.)}

\item{max.age}{Maximum age of the oldest stratigraphic interval}

\item{strata}{Number of stratigraphic interval}

\item{return.useful}{If TRUE return the branch labels used to define taxon types}
}
\value{
dataframe of per interval taxon types.
}
\description{
Categorise per interval taxon types using the three-timer approach (Alroy, 2008)
}
\details{
Taxa types are detailed in Alroy (2008) and Alroy (2014) \cr
1. taxa sampled at all in a focal bin (Ns) \cr
2. taxa sampled in a bin but not immediately before or after (one-timers, or 1t) \cr
3a. taxa sampled immediately before and within the ith bin (two-timers, or 2ti) * referred to as two_t_a \cr
3b. or within and immediately after the ith bin (2ti+1) * referred to as two_t_b \cr
4. taxa sampled in three consecutive bins (three-timers, or 3t) \cr
5. taxa sampled before and after but not within a bin (part-timers, or Pt) \cr
n.b. all three timers are also two timers \cr
}
\examples{
# simulate tree & fossils
t = ape::rtree(20)
# budding speciation
sp = FossilSim::sim.taxonomy(t)
# simulate fossils
max = FossilSim::tree.max(t)
f = FossilSim::sim.fossils.intervals(taxonomy = sp, max.age = max, strata = 10, probabilities = rep(0.5,10), use.exact.times = FALSE)
# add extant occurrences
f = FossilSim::sim.extant.samples(f, taxonomy = sp)
# categorise interval types
interval.types.3t(f, max, 10)
}
\references{
Alroy, J et al. 2008. Dynamics of origination and extinction in the marine fossil record. PNAS 105: 11536-11542.
}
