% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.calculations.tree.wide.R
\name{interval.types.bc}
\alias{interval.types.bc}
\title{Categorise per interval taxon types using boundary crosser categories for a set of equal length intervals}
\usage{
interval.types.bc(fossils, max.age, strata, return.useful = FALSE)
}
\arguments{
\item{fossils}{Dataframe of sampled fossils (sp = unique species ID. h = ages.)}

\item{max.age}{Maximum age of the oldest stratigraphic interval}

\item{strata}{Number of stratigraphic intervals}

\item{return.useful}{If TRUE return the branch labels used to define taxon types}
}
\value{
dataframe of per interval taxon types.
}
\description{
Categorise per interval taxon types using boundary crosser categories for a set of equal length intervals
}
\details{
Taxa types are detailed in Foote (2000). \cr
Nbt - bottom and top boundary crossers \cr
Nbl - bottom boundary crossers only (e.g. species goes extinct) \cr
NFt - top boundary crossers only (e.g. species originates) \cr
NFl - species originates and becomes extinct during the same interval \cr
}
\examples{
# simulate tree & fossils
t = ape::rtree(20)
# budding speciation
sp = FossilSim::sim.taxonomy(t)
# simulate fossils
max = FossilSim::tree.max(t)
f = FossilSim::sim.fossils.intervals(taxonomy = sp, max.age = max, strata = 10, probabilities = rep(0.5,10), use.exact.times = FALSE)
# add extant occurrences
f = FossilSim::sim.extant.samples(f, taxonomy = sp)
# categorise interval types
interval.types.bc(f, max, 10)
}
\references{
Foote, M. 2000. Origination and extinction components of taxonomic diversity: General problems. Paleobiology 26: 74-102.
}
