% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.calculations.tree.wide.R
\name{interval.types.gf}
\alias{interval.types.gf}
\title{Categorise per interval taxon types using the gap-filler approach (Alroy, 2014)}
\usage{
interval.types.gf(fossils, max.age, strata, return.useful = FALSE)
}
\arguments{
\item{fossils}{Dataframe of sampled fossils (sp = unique species ID. h = ages.)}

\item{max.age}{Maximum age of the oldest stratigraphic interval}

\item{strata}{Number of stratigraphic intervals}

\item{return.useful}{If TRUE return the branch labels used to define taxon types}
}
\value{
dataframe of per interval taxon types.
}
\description{
Categorise per interval taxon types using the gap-filler approach (Alroy, 2014)
}
\details{
Taxa types are detailed in Alroy (2014) \cr
1. taxa sampled at all in the focal bin (Ns) \cr
2. taxa sampled in a bin but not immediately before or after (one-timers, or 1t) \cr
3a. taxa sampled immediately before and within the ith bin (two-timers, or 2ti) * referred to as two_t_a \cr
3b. or within and immediately after the ith bin (2ti+1) * referred to as two_t_b \cr
4. taxa sampled in three consecutive bins (three-timers, or 3t) \cr
5. taxa sampled before and after but not within a bin (part-timers, or Pt) \cr
6. taxa sampled in i − 1 and i + 2 but not i + 1 (gap-fillers, a) * referred to as gf_a # may also be sampled in i \cr
7. taxa sampled in i + 1 and i - 2 but not i - 1 (gap-fillers, b) * referred to as gf_b # may also be sampled in i \cr
}
\examples{
# simulate tree & fossils
t = ape::rtree(20)
# budding speciation
sp = FossilSim::sim.taxonomy(t)
# simulate fossils
max = FossilSim::tree.max(t)
f = FossilSim::sim.fossils.intervals(taxonomy = sp, max.age = max, strata = 10, probabilities = rep(0.5,10), use.exact.times = FALSE)
# add extant occurrences
f = FossilSim::sim.extant.samples(f, taxonomy = sp)
# categorise interval types
interval.types.gf(f, max, 10)
}
\references{
Alroy, J 2014. Accurate and precise estimates of origination and extinction rates. Paleobiology 40: 374-397.
}
