% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.calculations.empirical.R
\name{presence.absence.matrix}
\alias{presence.absence.matrix}
\title{Generate presence/absence matrix from simulated data}
\usage{
presence.absence.matrix(fossils, intervals, add.extant = TRUE)
}
\arguments{
\item{fossils}{Fossil dataframe.}

\item{intervals}{Intervals dataframe.}
}
\value{
Dataframe recording presences & absences during different time intervals.
}
\description{
Generate presence/absence matrix from simulated data
}
\examples{
set.seed(123)
# simulate tree
t = TreeSim::sim.bd.taxa(100,1,0.01,0.005)[[1]]

# simulate species
sp = FossilSim::sim.taxonomy(t)

# simulate fossils
f = FossilSim::sim.fossils.poisson(1, taxonomy = sp)
f = FossilSim::sim.extant.samples(f, taxonomy = sp)

# generate presence.absence matrix
presence.absence.matrix(f, stages)

}
