% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{presence.absence.precooked}
\alias{presence.absence.precooked}
\title{Precooked presence absence matrix}
\format{A data frame with 1/0 data for each taxon (row) during each interval (columns).}
\usage{
presence.absence.precooked
}
\description{
A table indicating the presence/absence of taxa during different Phanerozoic stages.
}
\keyword{datasets}

