% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stages}
\alias{stages}
\title{Geological stages}
\format{A data frame with Phanerozoic stage start an end times:
\describe{
\item{stage}{stage name}
\item{start}{start of stage in MA}
\item{end}{end of stage in MA}
}}
\usage{
stages
}
\description{
A table containing the start and end times of the geological stages.
}
\keyword{datasets}
