% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.calculations.R
\name{three.timer.rates}
\alias{three.timer.rates}
\title{Calculate speciation and extinction rates using the three-timer approach}
\usage{
three.timer.rates(fossils, basin.age, strata, continuous = T)
}
\arguments{
\item{fossils}{Dataframe of sampled fossils (sp = edge labels. h = ages.)}

\item{basin.age}{Maximum age of the oldest stratigraphic interval}

\item{strata}{Number of stratigraphic horizons}

\item{continuous}{If TRUE calculate continuous rates}
}
\description{
Calculate speciation and extinction rates using the three-timer approach
}

