% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd.calculations.R
\name{bd.probability.extant}
\alias{bd.probability.extant}
\title{Birth-death probability (Stadler, 2012)}
\usage{
bd.probability.extant(tree, b, d, rho = 1, mpfr = FALSE, crown = FALSE)
}
\arguments{
\item{tree}{Phylo object of extant taxa (the function will remove any extinct taxa prior to calculating the likelihood)}

\item{b}{Rate of speciation (branching)}

\item{d}{Rate of extinction (branch termination). Must be < b}

\item{rho}{Extant species sampling probability}

\item{crown}{If TRUE the process is conditioned on the crown (default = F)}
}
\value{
Log likelihood
}
\description{
Probability of extant species phylogeny conditioned on the origin (Stadler, 2012, eq. 2)
or the crown (crown = T) (Stadler, 2012, eq. 5)
}
\examples{
# simulate tree
t = TreeSim::sim.bd.taxa(100,1,1,0.1)[[1]]
# calculate birth-death probability
birth = 1
death = 0.1
bd.probability.extant(t, birth, death)
}
\references{
Stadler, T. 2012. How can we improve accuracy of macroevolutionary rate estimates? Systematic Biology 62: 321-329.
}
