% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd.calculations.R
\name{bd.probability.range}
\alias{bd.probability.range}
\title{Birth-death probability (Keiding, 1975)}
\usage{
bd.probability.range(frs, b, d, crown = FALSE)
}
\arguments{
\item{frs}{Dataframe of species ranges}

\item{b}{Rate of speciation}

\item{d}{Rate of extinction}

\item{crown}{If TRUE assume the process begins at the first speciation event and not the origin (default = FALSE)}
}
\value{
Log likelihood
}
\description{
Calculate the probabilty of speciaton and extinction rates for a given set of stratigraphic ranges
}
\examples{
# simulate tree & assume complete sampling
t = TreeSim::sim.bd.taxa(100,1,1,0.1)[[1]]
# add symmetric speciation events & generate fossil range dataframe
f = 0.5
ages <- FossilSim::mixed.ages(t, f, root.edge = TRUE)
# add anagenic speciation events
lambda.a = 0.1
frs <- FossilSim::anagenic.species(ages, lambda.a)
# calculate birth-death probability
birth = 1
death = 0.1
bd.probability.range(frs, birth, death)
}
\references{
Keiding, N. 1975. Maximum likelihood estimation in the birth-death process. Annals of Statistics 3: 363-372.
}
