% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd.calculations.R
\name{bd.probability.range}
\alias{bd.probability.range}
\title{Birth-death probability (Keiding, 1975)}
\usage{
bd.probability.range(frs, b, d, crown = FALSE)
}
\arguments{
\item{frs}{Dataframe of species ranges}

\item{b}{Rate of speciation}

\item{d}{Rate of extinction}
}
\value{
Log likelihood
}
\description{
This is the underlying birth-death process used in Silvestro et al. 2014, eq. 9
}

