% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.calculations.R
\name{boundary.crosser.rates}
\alias{boundary.crosser.rates}
\title{Calculate speciation and extinction rates using the boundary crosser approach}
\usage{
boundary.crosser.rates(taxa.types, basin.age, strata, instantaneous = F)
}
\arguments{
\item{taxa.types}{Dataframe of per interval taxa types}

\item{basin.age}{Maximum age of the oldest stratigraphic interval}

\item{instantaneous}{If TRUE calculate instantaneous rates}

\item{Number}{of stratigraphic horizons}
}
\description{
Calculate speciation and extinction rates using the boundary crosser approach
}

