% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.calculations.R
\name{boundary.crosser.rates}
\alias{boundary.crosser.rates}
\title{Calculate speciation and extinction rates using the boundary crosser approach}
\usage{
boundary.crosser.rates(fossils, basin.age, strata, continuous = T,
  return.intervals = T)
}
\arguments{
\item{fossils}{Dataframe of sampled fossils (sp = edge labels. h = ages.)}

\item{basin.age}{Maximum age of the oldest stratigraphic interval}

\item{strata}{Number of stratigraphic horizons}

\item{continuous}{If TRUE calculate continuous rates}

\item{return.intervals}{If TRUE return per interval estimates}
}
\value{
named list with the overall speciation rate, overall extinction rate and a dataframe of per interval estimtes if return.intervals = TRUE.
}
\description{
Calculate speciation and extinction rates using the boundary crosser approach
}

