% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.calculations.R
\name{boundary.crosser.rates}
\alias{boundary.crosser.rates}
\title{Calculate speciation and extinction rates using the boundary crosser approach}
\usage{
boundary.crosser.rates(fossils, basin.age, strata, continuous = T,
  return.intervals = F)
}
\arguments{
\item{fossils}{Dataframe of sampled fossils (sp = edge labels. h = ages.)}

\item{basin.age}{Maximum age of the oldest stratigraphic interval}

\item{strata}{Number of stratigraphic intervals}

\item{continuous}{If TRUE calculate continuous rates}

\item{return.intervals}{If TRUE return per interval estimates}
}
\value{
named list with the overall speciation rate, overall extinction rate and a dataframe of per interval estimtes if return.intervals = TRUE.
Note this approach does not return rates for the first interval.
}
\description{
Calculate speciation and extinction rates using the boundary crosser approach
}
\examples{
# simulate tree & fossils
t = TreeSim::sim.bd.taxa(100,1,1,0.1)[[1]]
ba = FossilSim::basin.age(t)
f = FossilSim::sim.fossils.unif(t, ba, 20, 0.5)
# add extant occurrences
f = FossilSim::add.extant.occ(t, f)
# asymmetric taxon mapping
f = FossilSim::asymmetric.fossil.mapping(t, f)
# categorise interval types
boundary.crosser.rates(f, ba, 20)
}

