% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.calculations.R
\name{first.last.appearances}
\alias{first.last.appearances}
\title{Calculate out first and last appearances}
\usage{
first.last.appearances(fossils)
}
\arguments{
\item{fossils}{Dataframe of sampled fossils (sp = edge labels. h = ages.)}
}
\value{
Function returns the first and last appearances for each lineage, represented by the max secure age of the sampling horizon.
}
\description{
Calculate out first and last appearances
}

