% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.calculations.R
\name{gap.filler.rates}
\alias{gap.filler.rates}
\title{Calculate speciation and extinction rates using the gap-filler approach}
\usage{
gap.filler.rates(taxa.types, basin.age, strata)
}
\arguments{
\item{taxa.types}{Dataframe of per interval taxa types}

\item{basin.age}{Maximum age of the oldest stratigraphic interval}

\item{Number}{of stratigraphic horizons}

\item{instantaneous}{If TRUE calculate instantaneous rates}
}
\description{
Calculate speciation and extinction rates using the gap-filler approach
}

