% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.calculations.tree.wide.R
\name{gap.filler.rates}
\alias{gap.filler.rates}
\title{Calculate speciation and extinction rates using the gap-filler approach}
\usage{
gap.filler.rates(fossils, max.age, strata, return.intervals = FALSE)
}
\arguments{
\item{fossils}{Dataframe of sampled fossils (sp = unique species ID. h = ages.)}

\item{strata}{Number of stratigraphic intervals}

\item{return.intervals}{If TRUE return per interval estimates}

\item{basin.age}{Maximum age of the oldest stratigraphic interval}
}
\value{
named list with the overall speciation rate, overall extinction rate and a dataframe of per interval estimtes if return.intervals = TRUE.
Note this approach does not return specation rates for the first or second interval and extinction rate for the first or last interval.
Function accounts for interval length. (i.e. continuous = TRUE is hard coded)
}
\description{
Calculate speciation and extinction rates using the gap-filler approach
}
\examples{
# simulate tree & fossils
t = TreeSim::sim.bd.taxa(100,1,1,0.1)[[1]]
# budding speciation
sp = FossilSim::sim.taxonomy(t)
# simulate fossils
max = FossilSim::tree.max(t)
f = FossilSim::sim.fossils.intervals(taxonomy = sp, max.age = max, strata = 10, probabilities = rep(0.5,10), use.exact.times = FALSE)
# add extant occurrences
f = FossilSim::sim.extant.samples(f, taxonomy = sp)
# calculate speciation and extinction rates
gap.filler.rates(f, max, 10)
}
\references{
Alroy, J 2014. Accurate and precise estimates of origination and extinction rates. Paleobiology 40: 374-397.
}
