% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.calculations.R
\name{interval.types.3t}
\alias{interval.types.3t}
\title{Categorize per interval taxon types using the three-timer approach (Alroy, 2008)}
\usage{
interval.types.3t(fossils, basin.age, strata)
}
\arguments{
\item{fossils}{Dataframe of sampled fossils (sp = edge labels. h = ages.)}

\item{basin.age}{Maximum age of the oldest stratigraphic interval}

\item{strata}{Number of stratigraphic horizons}
}
\value{
dataframe of per interval taxon types.
}
\description{
Categorize per interval taxon types using the three-timer approach (Alroy, 2008)
}
\details{
Taxa types are detailed in Alroy (2000) and Alroy (2014) \cr
1. taxa sampled at all in a focal bin (Ns)
2. taxa sampled in a bin but not immediately before or after (one-timers, or 1t) \cr
3a. taxa sampled immediately before and within the ith bin (two-timers, or 2ti) * referred to as two_t_a \cr
3b. or within and immediately after the ith bin (2ti+1) * referred to as two_t_b \cr
4. taxa sampled in three consecutive bins (three-timers, or 3t) \cr
5. taxa sampled before and after but not within a bin (part-timers, or Pt) \cr
n.b. all three timers are also two timers \cr
}

