% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.calculations.R
\name{interval.types.3t}
\alias{interval.types.3t}
\title{Categorise per interval taxon types using the three-timer approach (Alroy, 2008)}
\usage{
interval.types.3t(fossils, basin.age, strata, return.useful = FALSE)
}
\arguments{
\item{fossils}{Dataframe of sampled fossils (sp = edge labels. h = ages.)}

\item{basin.age}{Maximum age of the oldest stratigraphic interval}

\item{strata}{Number of stratigraphic interval}

\item{return.useful}{If TRUE return the branch labels used to define taxon types}
}
\value{
dataframe of per interval taxon types.
}
\description{
Categorise per interval taxon types using the three-timer approach (Alroy, 2008)
}
\details{
Taxa types are detailed in Alroy (2008) and Alroy (2014) \cr
1. taxa sampled at all in a focal bin (Ns) \cr
2. taxa sampled in a bin but not immediately before or after (one-timers, or 1t) \cr
3a. taxa sampled immediately before and within the ith bin (two-timers, or 2ti) * referred to as two_t_a \cr
3b. or within and immediately after the ith bin (2ti+1) * referred to as two_t_b \cr
4. taxa sampled in three consecutive bins (three-timers, or 3t) \cr
5. taxa sampled before and after but not within a bin (part-timers, or Pt) \cr
n.b. all three timers are also two timers \cr
}
\examples{
# simulate tree & fossils
t = ape::rtree(10)
ba = FossilSim::basin.age(t)
f = FossilSim::sim.fossils.unif(t, ba, 10, 0.5)
# add extant occurrences
f = FossilSim::add.extant.occ(t, f)
# asymmetric taxon mapping
f = FossilSim::asymmetric.fossil.mapping(t, f)
# categorise interval types
interval.types.3t(f, ba, 10)
}
\references{
Alroy, J et al. 2008. Dynamics of origination and extinction in the marine fossil record. PNAS 105: 11536-11542.
}
