% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.calculations.empirical.R
\name{interval.types.bc.pa}
\alias{interval.types.bc.pa}
\title{Categorise per interval taxon types using boundary crosser categories based on presence/absence data}
\usage{
interval.types.bc.pa(presence.absence, intervals)
}
\arguments{
\item{presence.absence}{Dataframe of presence/absence data. Each row represents a taxon and each column represents a time interval from oldest to youngest.}

\item{intervals}{Dataframe of geological intervals.}
}
\value{
dataframe of per interval taxon types.
}
\description{
Categorise per interval taxon types using boundary crosser categories based on presence/absence data
}
\details{
Taxa types are detailed in Foote (2000). \cr
Nbt - bottom and top boundary crossers \cr
Nbl - bottom boundary crossers only (e.g. species goes extinct) \cr
NFt - top boundary crossers only (e.g. species originates) \cr
NFl - species originates and becomes extinct during the same interval \cr
}
\examples{
set.seed(123)
# simulate tree
t = TreeSim::sim.bd.taxa(100,1,0.01,0.005)[[1]]

# simulate species
sp = FossilSim::sim.taxonomy(t)

# simulate fossils
f = FossilSim::sim.fossils.poisson(1, taxonomy = sp)
f = FossilSim::sim.extant.samples(f, taxonomy = sp)

# generate presence.absence matrix
pa = presence.absence.matrix(f, stages)

# categorise interval types
interval.types.bc.pa(pa, stages)

}
\references{
Foote, M. 2000. Origination and extinction components of taxonomic diversity: General problems. Paleobiology 26: 74-102.
}
