% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.calculations.R
\name{interval.types.bc}
\alias{interval.types.bc}
\title{Categorize per interval taxon types using the boundary crosser approach}
\usage{
interval.types.bc(fossils, basin.age, strata)
}
\arguments{
\item{fossils}{Dataframe of sampled fossils (sp = edge labels. h = ages.)}

\item{basin.age}{Maximum age of the oldest stratigraphic interval}

\item{strata}{Number of stratigraphic horizons}
}
\value{
dataframe of per interval taxon types.
}
\description{
Categorize per interval taxon types using the boundary crosser approach
}
\details{
Taxa types are detailed in Foote (2000). \cr
Nbt - bottom and top boundary crossers \cr
Nbl - bottom boundary crossers only (e.g. species goes extinct) \cr
NFt - top boundary crossers only (e.g. species originates) \cr
NFl - species originates and becomes extinct during the same interval \cr
}

