% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.calculations.R
\name{interval.types.gf}
\alias{interval.types.gf}
\title{Categorize per interval taxon types using the gap-filler approach (Alroy, 2014)}
\usage{
interval.types.gf(fossils, basin.age, strata, return.useful = FALSE)
}
\arguments{
\item{fossils}{Dataframe of sampled fossils (sp = edge labels. h = ages.)}

\item{basin.age}{Maximum age of the oldest stratigraphic interval}

\item{strata}{Number of stratigraphic horizons}

\item{return.useful}{If TRUE return the branch labels used to define taxon types}
}
\value{
dataframe of per interval taxon types.
}
\description{
Categorize per interval taxon types using the gap-filler approach (Alroy, 2014)
}
\details{
Taxa types are detailed in Alroy (2014) \cr
1. taxa sampled at all in the focal bin (Ns) \cr
2. taxa sampled in a bin but not immediately before or after (one-timers, or 1t) \cr
3a. taxa sampled immediately before and within the ith bin (two-timers, or 2ti) * referred to as two_t_a \cr
3b. or within and immediately after the ith bin (2ti+1) * referred to as two_t_b \cr
4. taxa sampled in three consecutive bins (three-timers, or 3t) \cr
5. taxa sampled before and after but not within a bin (part-timers, or Pt) \cr
6. taxa sampled in i − 1 and i + 2 but not i + 1 (gap-fillers, a) * referred to as gf_a # may also be sampled in i \cr
7. taxa sampled in i + 1 and i - 2 but not i - 1 (gap-fillers, b) * referred to as gf_b # may also be sampled in i \cr
}

