% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.calculations.pbdb.R
\name{presence.absence.matrix}
\alias{presence.absence.matrix}
\title{Generate presence/absence matrix from simulated data}
\usage{
presence.absence.matrix(fossils, intervals, add.extant = TRUE)
}
\arguments{
\item{fossils}{Fossil dataframe.}

\item{intervals}{Intervals dataframe.}
}
\value{
Dataframe recording presences & absences during different time intervals.
}
\description{
Generate presence/absence matrix from simulated data
}
\examples{
set.seed(123)
# simulate tree
t = TreeSim::sim.bd.taxa(100,1,0.01,0.005)[[1]]

# simulate fossils
f = FossilSim::sim.fossils.poisson(t, 1)

# asymmetric species assignment
f = FossilSim::asymmetric.fossil.mapping(t, f)

# generate presence.absence matrix
presence.absence.matrix(f, stages)

}

