% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.calculations.R
\name{uncorrected.rates}
\alias{uncorrected.rates}
\title{Calculate speciation and extinction rates using the uncorrected approach}
\usage{
uncorrected.rates(taxa.types, basin.age, strata, instantaneous = F)
}
\arguments{
\item{taxa.types}{Dataframe of per interval taxa types}

\item{basin.age}{Maximum age of the oldest stratigraphic interval}

\item{instantaneous}{If TRUE calculate instantaneous rates}

\item{Number}{of stratigraphic horizons}
}
\description{
Calculate speciation and extinction rates using the uncorrected approach
}

