% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.calculations.tree.wide.R
\name{uncorrected.rates}
\alias{uncorrected.rates}
\title{Calculate speciation and extinction rates using the uncorrected approach}
\usage{
uncorrected.rates(fossils, max.age, strata, continuous = TRUE,
  return.intervals = FALSE)
}
\arguments{
\item{fossils}{Dataframe of sampled fossils (sp = unique species ID. h = ages.)}

\item{max.age}{Maximum age of the oldest stratigraphic interval}

\item{strata}{Number of stratigraphic intervals}

\item{continuous}{If TRUE calculate continuous rates (i.e. account for interval length)}

\item{return.intervals}{If TRUE return per interval estimates}
}
\value{
named list with the overall speciation rate, overall extinction rate and a dataframe of per interval estimtes if return.intervals = TRUE.
}
\description{
Calculate speciation and extinction rates using the uncorrected approach
}
\examples{
# simulate tree & fossils
t = TreeSim::sim.bd.taxa(100,1,1,0.1)[[1]]
# budding speciation
sp = FossilSim::sim.taxonomy(t)
# simulate fossils
max = FossilSim::tree.max(t)
f = FossilSim::sim.fossils.intervals(taxonomy = sp, max.age = max, strata = 10, probabilities = rep(0.5,10), use.exact.times = FALSE)
# add extant occurrences
f = FossilSim::sim.extant.samples(f, taxonomy = sp)
# calculate speciation and extinction rates
uncorrected.rates(f, max, 10)
}
